Rename variables, changed dependency
