+++ Changefeed/ChangefeedSample/ChangefeedSample.csproj
@@ -4,7 +4,7 @@
     <AzureFunctionsVersion>v3</AzureFunctionsVersion>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Azure.Storage.Blobs.ChangeFeed" Version="12.0.0-dev.20200731.1" />
+    <PackageReference Include="Azure.Storage.Blobs.Changefeed" Version="12.0.0-preview.2" />
     <PackageReference Include="Microsoft.NET.Sdk.Functions" Version="3.0.3" />
   </ItemGroup>
   <ItemGroup>
