+++ Changefeed/ChangefeedSample/ChangefeedSample.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30309.148
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ChangefeedSample", "ChangefeedSample.csproj", "{B724C481-C20D-4C9B-BB57-704DB4917691}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B724C481-C20D-4C9B-BB57-704DB4917691}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B724C481-C20D-4C9B-BB57-704DB4917691}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B724C481-C20D-4C9B-BB57-704DB4917691}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B724C481-C20D-4C9B-BB57-704DB4917691}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A86C585F-E030-43CB-A542-EB52457E1234}
+	EndGlobalSection
+EndGlobal
