--- Changefeed/ChangefeedSample/changefeedjd.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30204.135
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ChangefeedSample", "changefeedjd\ChangefeedSample.csproj", "{C51B7F4B-2AA6-4ED6-9103-46EF4343BF90}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C51B7F4B-2AA6-4ED6-9103-46EF4343BF90}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C51B7F4B-2AA6-4ED6-9103-46EF4343BF90}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C51B7F4B-2AA6-4ED6-9103-46EF4343BF90}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C51B7F4B-2AA6-4ED6-9103-46EF4343BF90}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {274655D7-75E2-4F2B-BBBB-98304FE2AD69}
-	EndGlobalSection
-EndGlobal
